Install os-*-config tools on nodes
These tools are required for making post-boot changes. Typical
use-case is scale-up/down of openshift nodes. Scripts are taken
from heat-templates repository [1] and slightly updated for usage
on centos. We might eventually refer directly to heat-templates
when including os-*-tools in templates but because some
centos-specific tweaks are required for now and set of required files
is not so big this was most straightforward way.