Check if os-collect-config is present during cloud-init
If cloud-init uses multipart MIME, then all scripts run no matter if
previous in chain failed, so it's possible that false-positive success is
reported. If os-collect-config is not started, then none of post-deployment
scripts is executed and stack-create times out. It also may cause that
stack-delete is stuck in DELETE_IN_PROGRESS because heat waits on
executing pre-delete scripts which are done by os-collect-config on
master node. It would be better to check if os-collect-config is running
but the problem is that os-collect-config is started once cloud-init (which
is systemd service too) finishes.