Add skip_dns parameter
If skip_dns is set to True, then DNS server is not set on
infrastructure node and DNS records are not managed for openshift
nodes. This is useful when using an external DNS server or DNSaaS
(e.g. Designate). Then user is expected to take care of managing
DNS records for nova VMs.