Remove node's contraints
Any cotnraint causes that heat does an external API call when
validating the stack (on any update/create operation). This is a big
problem for large deployments, e.g. 6 contraints took ~6 seconds
per single node -> with 50 nodes it took 5 mins to only validate because
validation is sequential.

These params are validated in the top-level openshift.yaml templates
so it's safe to remove them to speed things up.